Drop macOS support
Macs can run the iPad version of the game.
