--- Shotty Bird macOS/AppDelegate.swift
@@ -1,32 +1,0 @@
-//
-//  AppDelegate.swift
-//  Shotty Bird macOS
-//
-//  Copyright © 2023 Komodo Life. All rights reserved.
-//
-
-import Cocoa
-
-@main
-class AppDelegate: NSObject, NSApplicationDelegate {
-
-
-
-    func applicationDidFinishLaunching(_ aNotification: Notification) {
-        // Insert code here to initialize your application
-        if let window = NSApplication.shared.windows.last {
-            window.toggleFullScreen(nil)
-        }
-    }
-
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
-    }
-
-    func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
-        return true
-    }
-
-
-}
-
