--- Shotty Bird macOS/GameViewController.swift
@@ -1,30 +1,0 @@
-//
-//  GameViewController.swift
-//  Shotty Bird macOS
-//
-//  Copyright © 2023 Komodo Life. All rights reserved.
-//
-
-import Cocoa
-import SpriteKit
-import GameplayKit
-
-class GameViewController: NSViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        let scene = MainMenuScene(backgroundSpeed: .slow)
-        
-        // Present the scene
-        let skView = self.view as! SKView
-        skView.presentScene(scene)
-        
-        skView.ignoresSiblingOrder = true
-        
-        skView.showsFPS = true
-        skView.showsNodeCount = true
-    }
-
-}
-
